+++ src/keyboard/keyboard.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-uint32_t keyboard_scan_key();
+uint8_t keyboard_scan_key();
