+++ src/memory/heap/heap.c
@@ -1,12 +1,12 @@
 #include "heap.h"
-#include "status.h"
 #include "memory/memory.h"
-#include "terminal/terminal.h"
-#include "whitelib/string.h"
+#include "status.h"
+#include <stdint.h>
 
-//static numstr_buff[32];
+#include "config.h"
 
-static int heap_validate_table(void* ptr, void* end, struct heap_table* table) {
+static int heap_validate_table(void* ptr, void* end, struct heap_table* table)
+{
     int res = 0;
 
     size_t table_size = (size_t)(end - ptr);
@@ -21,11 +21,13 @@ out:
     return res;
 }
 
-static int heap_validate_alignment(void* ptr) {
-    return ((unsigned int)ptr % WHITEOS_HEAP_BLOCK_SIZE) == 0;
+static int heap_validate_alignment(void* ptr)
+{
+    return ((intptr_t)ptr % WHITEOS_HEAP_BLOCK_SIZE) == 0;
 }
 
-int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* table) {
+int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* table)
+{
     int res = 0;
 
     if (!heap_validate_alignment(ptr) || !heap_validate_alignment(end))
@@ -52,7 +54,8 @@ out:
     return res;
 }
 
-static uint32_t heap_align_value_to_upper(uint32_t val) {
+static uint32_t heap_align_value_to_upper(uint32_t val)
+{
     //aligns val to multiple of BLOCK_SIZE
     return val + ((uint32_t)(val % WHITEOS_HEAP_BLOCK_SIZE != 0) * (WHITEOS_HEAP_BLOCK_SIZE - (val % WHITEOS_HEAP_BLOCK_SIZE)));
 }
@@ -61,7 +64,8 @@ static int heap_get_entry_type(HEAP_BLOCK_TABLE_ENTRY entry) {
     return entry & 0x0F;
 }
 
-int heap_get_start_block(struct heap* heap, uint32_t total_blocks) {
+int heap_get_start_block(struct heap* heap, uint32_t total_blocks)
+{
     struct heap_table* table = heap->table;
     int bc = 0;     // current block
     int bs = -1;    // start block
@@ -85,18 +89,20 @@ int heap_get_start_block(struct heap* heap, uint32_t total_blocks) {
             break;
         }
     }
-    if (bs == -1) 
+    if (bs == -1)
     {
         return -ENOMEM;
     }
     return bs;
 }
 
-void* heap_block_to_address(struct heap* heap, int block) {
+void* heap_block_to_address(struct heap* heap, int block)
+{
     return heap->saddr + (block * WHITEOS_HEAP_BLOCK_SIZE);
 }
 
-void heap_mark_blocks_taken(struct heap* heap, int start_block, int total_blocks) {
+void heap_mark_blocks_taken(struct heap* heap, int start_block, int total_blocks)
+{
     int end_block = start_block + total_blocks - 1;
 
     HEAP_BLOCK_TABLE_ENTRY entry = HEAP_BLOCK_TABLE_ENTRY_TAKEN | HEAP_BLOCK_IS_FIRST;
@@ -104,7 +110,7 @@ void heap_mark_blocks_taken(struct heap* heap, int start_block, int total_blocks
     {
         entry |= HEAP_BLOCK_HAS_NEXT;
     }
-    
+
     for (int i=start_block; i <= end_block; i++)
     {
         heap->table->entries[i] = entry;
@@ -116,7 +122,8 @@ void heap_mark_blocks_taken(struct heap* heap, int start_block, int total_blocks
     }
 }
 
-void* heap_malloc_blocks(struct heap* heap, uint32_t total_blocks) {
+void* heap_malloc_blocks(struct heap* heap, uint32_t total_blocks)
+{
     void* address = 0;
     int start_block = heap_get_start_block(heap, total_blocks);
 
@@ -132,11 +139,13 @@ out:
     return address;
 }
 
-int heap_address_to_block(struct heap* heap, void* address) {
+int heap_address_to_block(struct heap* heap, void* address)
+{
     return (int)(address - heap->saddr) / WHITEOS_HEAP_BLOCK_SIZE;
 }
 
-void heap_mark_blocks_free(struct heap* heap, int starting_block) {
+void heap_mark_blocks_free(struct heap* heap, int starting_block)
+{
     struct heap_table* table = heap->table;
 
     for (int i = starting_block; i < (int)table->total; i++)
@@ -151,14 +160,16 @@ void heap_mark_blocks_free(struct heap* heap, int starting_block) {
     }
 }
 
-void* heap_malloc(struct heap* heap, size_t size) {
+void* heap_malloc(struct heap* heap, size_t size)
+{
     size_t aligned_size = heap_align_value_to_upper(size);
     uint32_t total_blocks = aligned_size / WHITEOS_HEAP_BLOCK_SIZE;
 
     return heap_malloc_blocks(heap, total_blocks);
 }
 
-void heap_free(struct heap* heap, void* ptr) {
+void heap_free(struct heap* heap, void* ptr)
+{
     heap_mark_blocks_free(heap, heap_address_to_block(heap, ptr));
     return;
 }
